% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFactors.R
\name{cleanFactors}
\alias{cleanFactors}
\title{cleanFactors}
\usage{
cleanFactors(nsubjects = 300, nlatent = NA, nvars = NA, ndistract = NA,
  noiselevel = 0, garbage = FALSE)
}
\arguments{
\item{nlatent}{Number of latent factors. By default generates a random number of factors between 2 and 6.}

\item{nvars}{Number of measured variables (items) for each latent factor. By default generated a random number of variables for each factor. Assigning a value generates the same number of variables for each factor.}

\item{ndistract}{Number of variable (items) not loading on any factor.}

\item{noiselevel}{An integer to add noise to the structure matrix used to create the factor structure.}

\item{garbage}{A boolean argument used to generate uniform data with some level of noise.}

\item{nsubject}{Number of rows of data to simulate.}
}
\value{
A list of components.
   \item{sample}{The sample dataset.}
   \item{nlatent}{The number of latent factors used to generate the dataset.}
   \item{nvars}{A vector of the number of variables in each factor.}
   \item{ndistract}{The number of variables not in the factor structure.}
   \item{nsubjects}{The number of subjects in the dataset.}
   \item{noiselevel}{The amount of noise added to the factor structure.}
}
\description{
Make factor analysis practice data.
}
