% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{efa}
\alias{efa}
\title{efa}
\usage{
efa(data, nfactors = 2, items = c(1:ncol(data)), cutoff = 0.3)
}
\arguments{
\item{data}{A data frame with the data to be analyzed. May contain missing data (which will be deleted listwise).}

\item{nfactors}{The number of factors to return.}

\item{items}{A vector of column numbers to include in the analysis.}

\item{cutoff}{The cutoff for loading display in the factor analysis. 0.3 by default.}
}
\value{
Prints the output of efa using principal axis/oblimin + recommendations for item reduction.
}
\description{
A wrapper function for running exploratory factor analysis.
}
