% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{heuristics}
\alias{heuristics}
\title{heuristics}
\usage{
heuristics(data, items = c(1:ncol(data)))
}
\arguments{
\item{data}{A data frame with the data to be analyzed. May contain missing data (which will be deleted listwise).}

\item{items}{A vector of column numbers to include in the analysis.}
}
\value{
Prints the results of 3 heuristics + eigenvalues > 1 (for purposes of mockery) + scree plot to graphics device.
}
\description{
Summary output of heuristics before and after factor analysis.
}
