% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverbank.R
\name{riverbank}
\alias{riverbank}
\title{riverbank}
\usage{
riverbank(factorlist)
}
\arguments{
\item{factorlist}{A list of factors from \code{watershed()}}
}
\value{
A list of components
}
\description{
A function to create edges and node for riverplot.
}
