% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverbend.R
\name{riverbend}
\alias{riverbend}
\title{riverbend}
\usage{
riverbend(data, maxlatent = NA, minlatent = NA, cutoff = 0.4)
}
\arguments{
\item{data}{Data frame containing only variables to test.}

\item{maxlatent}{Maximum number of latent variables to test.}

\item{minlatent}{Minimum number of latent variables to test.}

\item{cutoff}{Cutoff value for leaving a variable out of that factor analysis.}
}
\value{
Draws a riverplot
}
\description{
A function for automatically generating factor analysis riverplots wiht a sequence of solutions.
}
