% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walt.histogram.R
\name{walt.histogram}
\alias{walt.histogram}
\title{walt.histogram}
\usage{
walt.histogram(df, column)
}
\arguments{
\item{df}{A dataframe what needs a violin plot.}

\item{column}{The name of the column to plot.}
}
\value{
A ggplot.
}
\description{
A function for drawing the composite histogram.
}
