% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walt.violinplot.R
\name{walt.violinplot}
\alias{walt.violinplot}
\title{walt.violinplot}
\usage{
walt.violinplot(df, column)
}
\arguments{
\item{df}{A dataframe what needs a violin plot.}

\item{column}{The name of the column to plot.}
}
\value{
A ggplot.
}
\description{
Function for drawing the composite violin plot.
}
