% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waltKMO.R
\name{waltKMO}
\alias{waltKMO}
\title{waltKMO}
\usage{
waltKMO(x)
}
\arguments{
\item{x}{a dataframe containing all of the columns to be included in an analysis.}
}
\value{
A list of components.
   \item{KMO}{The overall measure of sampling adequacy.}
   \item{MSA}{Measures of sampling adequacy for each item.}
}
\description{
waltKMO is a function for calculating Kaiser-Meyer using the Kaiser & Rice (1974) algorithm.
}
\section{Interpretation}{

Kaiser and Rice (1974) categorized KMO values using the following subjective appraisals.
\tabular{rl}{
   in the 0.90s \tab marvelous
   in the 0.80s \tab meritorious
   in the 0.70s \tab middling
   in the 0.60s \tab mediocre
   in the 0.50s \tab miserable
   below 0.50 \tab unacceptable
}
}

