% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waltplot.R
\name{waltplot}
\alias{waltplot}
\title{waltplot}
\usage{
waltplot(vars, smoothed, data)
}
\arguments{
\item{vars}{A vector of column names to be plotted.}

\item{smoothed}{A parameter of whether to use histogram (0) or violinplots (1) in the waltplot grid.}

\item{data}{The data frame.}
}
\description{
A function to grid all the waltplots.
}
