% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{watershed}
\alias{watershed}
\title{watershed}
\usage{
watershed(data, maxlatent = ncol(data), minlatent = 1, cutoff = 0.4)
}
\arguments{
\item{data}{Data frame containing only variables to test.}

\item{maxlatent}{Maximum number of latent variables to test.}

\item{minlatent}{Minimum number of latent variables to test.}

\item{cutoff}{Cutoff value for leaving a variable out of that factor analysis.}
}
\value{
A data frame with maximum factor loading identified for each variable.
}
\description{
A function for constructing a summary data frame for a series of exploratory factor analyses.
}
